log group name has to match lambda name
For anything previously deployed, this will be a weird breaking change.
`terraform apply` will probably fail to create the new log group because
it already exists (the Lambda would have created it). If this happens to
you, the solution is to use `terraform import` to link the correctly
named log group in your terraform project's state file. Alternatively,
you could delete the log group and let terraform create it, but this
will be a race against any Lambda invocations.